更新: 测试了鼠标事件的工作情况, 目前可以直接用鼠标绘制像素了

+++ AmbiRenderer/core/ControlManager.cpp
@@ -15,7 +15,7 @@ void ControlManager::SetKeyUp(KeyCode k) noexcept
 	cur_kb_states[unsigned(k)] = KEY_UP;
 }
 
-void ControlManager::SetMousePos(int x, int y) noexcept
+void ControlManager::SetMousePos(float x, float y) noexcept
 {
 	cur_mouse_pos.x = x;
 	cur_mouse_pos.y = y;
